# docs/swagger.yaml
info:
  contact: {}
paths:
  /:
    get:
      description: do ping
      responses:
        "200":
          description: '{"message":"hello world"}'
          schema:
            type: string
      summary: ping example
      tags:
      - example
  /problem/detail:
    get:
      parameters:
      - description: 问题标识
        in: query
        name: identity
        required: true
        type: string
      responses:
        "200":
          description: 'status":"success","data":{"id": 1, "identity": "", "title":
            "", "content": "", "total_num": 0, "problem_categories": []}}'
          schema:
            type: object
        "400":
          description: status":"error","error":"错误信息"}
          schema:
            type: object
      summary: 问题详情
      tags:
      - 问题
  /problems:
    get:
      parameters:
      - description: 分页，默认1
        in: query
        name: page
        type: integer
      - description: 分页大小，默认15
        in: query
        name: size
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 分类标识
        in: query
        name: category_identity
        type: string
      responses:
        "200":
          description: 'status":"success","data":{"count": 1, "list": []}}'
          schema:
            type: object
        "400":
          description: status":"error","error":"错误信息"}
          schema:
            type: object
      summary: 问题列表
      tags:
      - 问题
  /submits:
    get:
      parameters:
      - description: 问题标识
        in: query
        name: problem_identity
        type: string
      - description: 用户标识
        in: query
        name: user_identity
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      - description: 分页，默认1
        in: query
        name: page
        type: integer
      - description: 分页大小，默认15
        in: query
        name: size
        type: integer
      responses:
        "200":
          description: 'status":"success","data":{"count": 1, "list": []}}'
          schema:
            type: object
        "400":
          description: status":"error","error":"错误信息"}
          schema:
            type: object
      summary: 提交列表
      tags:
      - 提交
  /user/detail:
    get:
      parameters:
      - description: 用户标识
        in: query
        name: identity
        required: true
        type: string
      responses:
        "200":
          description: status":"success","data":{}}
          schema:
            type: object
        "400":
          description: status":"error","error":"错误信息"}
          schema:
            type: object
      summary: 用户详情
      tags:
      - 用户
  /user/login:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: name
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      responses:
        "200":
          description: status":"success","data":{}}
          schema:
            type: object
        "400":
          description: status":"error","error":"错误信息"}
          schema:
            type: object
      summary: 用户登陆
      tags:
      - 用户
  /user/register:
    post:
      parameters:
      - description: 用户名
        in: formData
        name: name
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      - description: 手机
        in: formData
        name: phone
        required: true
        type: string
      - description: 验证码
        in: formData
        name: code
        required: true
        type: string
      responses:
        "200":
          description: status":"success","data":{}}
          schema:
            type: object
        "400":
          description: status":"error","error":"错误信息"}
          schema:
            type: object
      summary: 用户注册
      tags:
      - 用户
  /user/sendmail:
    get:
      parameters:
      - description: 接收邮件地址
        in: query
        name: email
        required: true
        type: string
      responses:
        "200":
          description: status":"success","data":{}}
          schema:
            type: object
        "400":
          description: status":"error","error":"错误信息"}
          schema:
            type: object
      summary: 发送邮件
      tags:
      - 用户
swagger: "2.0"